requirements.txt
torch>=1.0.0
torchvision>=0.2.1
dominate>=2.4.0
numpy>=1.15.0
scipy>=1.2.0
scikit-learn>=0.20.0
pillow>=6.0.0
matplotlib>=3.0.0
tqdm>=4.30.0
pydicom>=2.1.0